// src/schema/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /**
   * A date string, such as 2007-12-03, compliant with the `full-date` format
   * outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for
   * representation of dates and times using the Gregorian calendar.
   */
  Date: { input: any; output: any; }
  /**
   * A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the
   * `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO
   * 8601 standard for representation of dates and times using the Gregorian calendar.
   */
  DateTime: { input: any; output: any; }
  /** A string used to identify an i18n locale */
  I18NLocaleCode: { input: any; output: any; }
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: any; output: any; }
  /** The `Upload` scalar type represents a file upload. */
  Upload: { input: any; output: any; }
};
// ...
export type IgloohomeChangelog = {
  __typename?: 'IgloohomeChangelog';
  AppVersion?: Maybe<Scalars['String']['output']>;
  Changes?: Maybe<Scalars['String']['output']>;
  ReleaseDate?: Maybe<Scalars['Date']['output']>;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  locale?: Maybe<Scalars['String']['output']>;
  localizations?: Maybe<IgloohomeChangelogRelationResponseCollection>;
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};


export type IgloohomeChangelogLocalizationsArgs = {
  filters?: InputMaybe<IgloohomeChangelogFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  publicationState?: InputMaybe<PublicationState>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
};

export type IgloohomeChangelogEntity = {
  __typename?: 'IgloohomeChangelogEntity';
  attributes?: Maybe<IgloohomeChangelog>;
  id?: Maybe<Scalars['ID']['output']>;
};

export type IgloohomeChangelogEntityResponse = {
  __typename?: 'IgloohomeChangelogEntityResponse';
  data?: Maybe<IgloohomeChangelogEntity>;
};
// ...
export type IgloohomeChangelogRelationResponseCollection = {
  __typename?: 'IgloohomeChangelogRelationResponseCollection';
  data: Array<IgloohomeChangelogEntity>;
};
// ...
export type IgloohomeFaqCategory = {
  __typename?: 'IgloohomeFaqCategory';
  Description?: Maybe<Scalars['String']['output']>;
  Name?: Maybe<Scalars['String']['output']>;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  locale?: Maybe<Scalars['String']['output']>;
  localizations?: Maybe<IgloohomeFaqCategoryRelationResponseCollection>;
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};


export type IgloohomeFaqCategoryLocalizationsArgs = {
  filters?: InputMaybe<IgloohomeFaqCategoryFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  publicationState?: InputMaybe<PublicationState>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
};

export type IgloohomeFaqCategoryEntity = {
  __typename?: 'IgloohomeFaqCategoryEntity';
  attributes?: Maybe<IgloohomeFaqCategory>;
  id?: Maybe<Scalars['ID']['output']>;
};

export type IgloohomeFaqCategoryEntityResponse = {
  __typename?: 'IgloohomeFaqCategoryEntityResponse';
  data?: Maybe<IgloohomeFaqCategoryEntity>;
};
// ...
export type IgloohomeFaqCategoryRelationResponseCollection = {
  __typename?: 'IgloohomeFaqCategoryRelationResponseCollection';
  data: Array<IgloohomeFaqCategoryEntity>;
};

export type IgloohomeFaqEntity = {
  __typename?: 'IgloohomeFaqEntity';
  attributes?: Maybe<IgloohomeFaq>;
  id?: Maybe<Scalars['ID']['output']>;
};

export type IgloohomeFaqEntityResponse = {
  __typename?: 'IgloohomeFaqEntityResponse';
  data?: Maybe<IgloohomeFaqEntity>;
};
// ...
export type Mutation = {
  __typename?: 'Mutation';
  /** Change user password. Confirm with the current password. */
  changePassword?: Maybe<UsersPermissionsLoginPayload>;
  createIgloohomeChangelog?: Maybe<IgloohomeChangelogEntityResponse>;
  createIgloohomeChangelogLocalization?: Maybe<IgloohomeChangelogEntityResponse>;
  createIgloohomeFaq?: Maybe<IgloohomeFaqEntityResponse>;
  createIgloohomeFaqCategory?: Maybe<IgloohomeFaqCategoryEntityResponse>;
  createIgloohomeFaqCategoryLocalization?: Maybe<IgloohomeFaqCategoryEntityResponse>;
  createIgloohomeFaqLocalization?: Maybe<IgloohomeFaqEntityResponse>;
  createUploadFile?: Maybe<UploadFileEntityResponse>;
  createUploadFolder?: Maybe<UploadFolderEntityResponse>;
  /** Create a new role */
  createUsersPermissionsRole?: Maybe<UsersPermissionsCreateRolePayload>;
  /** Create a new user */
  createUsersPermissionsUser: UsersPermissionsUserEntityResponse;
  deleteIgloohomeChangelog?: Maybe<IgloohomeChangelogEntityResponse>;
  deleteIgloohomeFaq?: Maybe<IgloohomeFaqEntityResponse>;
  deleteIgloohomeFaqCategory?: Maybe<IgloohomeFaqCategoryEntityResponse>;
  deleteUploadFile?: Maybe<UploadFileEntityResponse>;
  deleteUploadFolder?: Maybe<UploadFolderEntityResponse>;
  /** Delete an existing role */
  deleteUsersPermissionsRole?: Maybe<UsersPermissionsDeleteRolePayload>;
  /** Delete an existing user */
  deleteUsersPermissionsUser: UsersPermissionsUserEntityResponse;
  /** Confirm an email users email address */
  emailConfirmation?: Maybe<UsersPermissionsLoginPayload>;
  /** Request a reset password token */
  forgotPassword?: Maybe<UsersPermissionsPasswordPayload>;
  login: UsersPermissionsLoginPayload;
  multipleUpload: Array<Maybe<UploadFileEntityResponse>>;
  /** Register a user */
  register: UsersPermissionsLoginPayload;
  removeFile?: Maybe<UploadFileEntityResponse>;
  /** Reset user password. Confirm with a code (resetToken from forgotPassword) */
  resetPassword?: Maybe<UsersPermissionsLoginPayload>;
  updateFileInfo: UploadFileEntityResponse;
  updateIgloohomeChangelog?: Maybe<IgloohomeChangelogEntityResponse>;
  updateIgloohomeFaq?: Maybe<IgloohomeFaqEntityResponse>;
  updateIgloohomeFaqCategory?: Maybe<IgloohomeFaqCategoryEntityResponse>;
  updateUploadFile?: Maybe<UploadFileEntityResponse>;
  updateUploadFolder?: Maybe<UploadFolderEntityResponse>;
  /** Update an existing role */
  updateUsersPermissionsRole?: Maybe<UsersPermissionsUpdateRolePayload>;
  /** Update an existing user */
  updateUsersPermissionsUser: UsersPermissionsUserEntityResponse;
  upload: UploadFileEntityResponse;
};
// ...
export type Query = {
  __typename?: 'Query';
  i18NLocale?: Maybe<I18NLocaleEntityResponse>;
  i18NLocales?: Maybe<I18NLocaleEntityResponseCollection>;
  igloohomeChangelog?: Maybe<IgloohomeChangelogEntityResponse>;
  igloohomeChangelogs?: Maybe<IgloohomeChangelogEntityResponseCollection>;
  igloohomeFaq?: Maybe<IgloohomeFaqEntityResponse>;
  igloohomeFaqCategories?: Maybe<IgloohomeFaqCategoryEntityResponseCollection>;
  igloohomeFaqCategory?: Maybe<IgloohomeFaqCategoryEntityResponse>;
  igloohomeFaqs?: Maybe<IgloohomeFaqEntityResponseCollection>;
  me?: Maybe<UsersPermissionsMe>;
  search?: Maybe<SearchResponse>;
  uploadFile?: Maybe<UploadFileEntityResponse>;
  uploadFiles?: Maybe<UploadFileEntityResponseCollection>;
  uploadFolder?: Maybe<UploadFolderEntityResponse>;
  uploadFolders?: Maybe<UploadFolderEntityResponseCollection>;
  usersPermissionsRole?: Maybe<UsersPermissionsRoleEntityResponse>;
  usersPermissionsRoles?: Maybe<UsersPermissionsRoleEntityResponseCollection>;
  usersPermissionsUser?: Maybe<UsersPermissionsUserEntityResponse>;
  usersPermissionsUsers?: Maybe<UsersPermissionsUserEntityResponseCollection>;
};
// ...
export type I18NLocalesQuery = { __typename?: 'Query', i18NLocales?: { __typename?: 'I18NLocaleEntityResponseCollection', data: Array<{ __typename?: 'I18NLocaleEntity', id?: string | null, attributes?: { __typename?: 'I18NLocale', code?: string | null } | null }> } | null };

export type CategoriesQueryVariables = Exact<{ [key: string]: never; }>;


export type CategoriesQuery = { __typename?: 'Query', igloohomeFaqCategories?: { __typename?: 'IgloohomeFaqCategoryEntityResponseCollection', data: Array<{ __typename?: 'IgloohomeFaqCategoryEntity', id?: string | null, attributes?: { __typename?: 'IgloohomeFaqCategory', Name?: string | null } | null }> } | null };

export type CategoryQueryVariables = Exact<{
  id?: InputMaybe<Scalars['ID']['input']>;
}>;


export type CategoryQuery = { __typename?: 'Query', igloohomeFaqCategory?: { __typename?: 'IgloohomeFaqCategoryEntityResponse', data?: { __typename?: 'IgloohomeFaqCategoryEntity', id?: string | null, attributes?: { __typename?: 'IgloohomeFaqCategory', Name?: string | null, publishedAt?: any | null, localizations?: { __typename?: 'IgloohomeFaqCategoryRelationResponseCollection', data: Array<{ __typename?: 'IgloohomeFaqCategoryEntity', id?: string | null, attributes?: { __typename?: 'IgloohomeFaqCategory', locale?: string | null } | null }> } | null } | null } | null } | null };

export type CreateIgloohomeFaqCategoryLocalizationMutationVariables = Exact<{
  locale?: InputMaybe<Scalars['I18NLocaleCode']['input']>;
  id?: InputMaybe<Scalars['ID']['input']>;
  data?: InputMaybe<IgloohomeFaqCategoryInput>;
}>;


export type CreateIgloohomeFaqCategoryLocalizationMutation = { __typename?: 'Mutation', createIgloohomeFaqCategoryLocalization?: { __typename?: 'IgloohomeFaqCategoryEntityResponse', data?: { __typename?: 'IgloohomeFaqCategoryEntity', id?: string | null, attributes?: { __typename?: 'IgloohomeFaqCategory', Name?: string | null } | null } | null } | null };

export type UpdateCategoryMutationVariables = Exact<{
  id: Scalars['ID']['input'];
  data: IgloohomeFaqCategoryInput;
}>;


export type UpdateCategoryMutation = { __typename?: 'Mutation', updateIgloohomeFaqCategory?: { __typename?: 'IgloohomeFaqCategoryEntityResponse', data?: { __typename?: 'IgloohomeFaqCategoryEntity', id?: string | null } | null } | null };
// ...
export type ChangelogQueryVariables = Exact<{
  id?: InputMaybe<Scalars['ID']['input']>;
}>;


export type ChangelogQuery = { __typename?: 'Query', igloohomeChangelog?: { __typename?: 'IgloohomeChangelogEntityResponse', data?: { __typename?: 'IgloohomeChangelogEntity', id?: string | null, attributes?: { __typename?: 'IgloohomeChangelog', AppVersion?: string | null, ReleaseDate?: any | null, Changes?: string | null, localizations?: { __typename?: 'IgloohomeChangelogRelationResponseCollection', data: Array<{ __typename?: 'IgloohomeChangelogEntity', id?: string | null, attributes?: { __typename?: 'IgloohomeChangelog', locale?: string | null } | null }> } | null } | null } | null } | null };

export type CreateIgloohomeChangelogLocalizationMutationVariables = Exact<{
  locale?: InputMaybe<Scalars['I18NLocaleCode']['input']>;
  id?: InputMaybe<Scalars['ID']['input']>;
  data?: InputMaybe<IgloohomeChangelogInput>;
}>;


export type CreateIgloohomeChangelogLocalizationMutation = { __typename?: 'Mutation', createIgloohomeChangelogLocalization?: { __typename?: 'IgloohomeChangelogEntityResponse', data?: { __typename?: 'IgloohomeChangelogEntity', id?: string | null } | null } | null };

export type UpdateChangelogMutationVariables = Exact<{
  id: Scalars['ID']['input'];
  data: IgloohomeChangelogInput;
}>;


export type UpdateChangelogMutation = { __typename?: 'Mutation', updateIgloohomeChangelog?: { __typename?: 'IgloohomeChangelogEntityResponse', data?: { __typename?: 'IgloohomeChangelogEntity', id?: string | null } | null } | null };
// ...
export type FaqQueryVariables = Exact<{
  id?: InputMaybe<Scalars['ID']['input']>;
}>;


export type FaqQuery = { __typename?: 'Query', igloohomeFaq?: { __typename?: 'IgloohomeFaqEntityResponse', data?: { __typename?: 'IgloohomeFaqEntity', id?: string | null, attributes?: { __typename?: 'IgloohomeFaq', Title?: string | null, localizations?: { __typename?: 'IgloohomeFaqRelationResponseCollection', data: Array<{ __typename?: 'IgloohomeFaqEntity', id?: string | null, attributes?: { __typename?: 'IgloohomeFaq', locale?: string | null } | null }> } | null } | null } | null } | null };

export type CreateIgloohomeFaqLocalizationMutationVariables = Exact<{
  locale?: InputMaybe<Scalars['I18NLocaleCode']['input']>;
  id?: InputMaybe<Scalars['ID']['input']>;
  data?: InputMaybe<IgloohomeFaqInput>;
}>;


export type CreateIgloohomeFaqLocalizationMutation = { __typename?: 'Mutation', createIgloohomeFaqLocalization?: { __typename?: 'IgloohomeFaqEntityResponse', data?: { __typename?: 'IgloohomeFaqEntity', id?: string | null } | null } | null };

export type UpdateIgloohomeFaqMutationVariables = Exact<{
  id: Scalars['ID']['input'];
  data: IgloohomeFaqInput;
}>;


export type UpdateIgloohomeFaqMutation = { __typename?: 'Mutation', updateIgloohomeFaq?: { __typename?: 'IgloohomeFaqEntityResponse', data?: { __typename?: 'IgloohomeFaqEntity', id?: string | null } | null } | null };